refactor(lessons): extraire useLessonEditor de LessonEditor (H4)
1191→989 lignes. Extraction du script (~231 l) dans composables/useLessonEditor.js
(256 l): état du formulaire, chargement leçon/matières/classes/chapitres, save
(création/édition), suppression, gestion des ressources. La vue passe en script setup
et se contente de câbler le composable.

DÉVIATION ASSUMÉE (choix explicite utilisateur): le template + CSS restent dans la
vue (pas d'éclatement des sections en sous-composants), donc LessonEditor.vue reste
à 989 lignes > 300. Cible ≤300 NON atteinte pour CE fichier, par décision produit
(éviter de fragmenter le formulaire / pas de partial CSS partagé). Dette tracée.

Parité: template + CSS inchangés verbatim; $router.go(-1) conservé via propriété
globale du routeur. Dette pré-existante conservée: CSS mort .loading-state/.spinner/
@keyframes spin (le template utilise <ContentLoader>).

Tests: composable 5 verts + test G5 de montage existant inchangé (1 vert).
